feat: reinforce role update security

+++ back-end/src/main/java/santos/tiago/back_end/controller/ApiController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import santos.tiago.back_end.model.User;
 import santos.tiago.back_end.model.UserResponse;
+import santos.tiago.back_end.model.UserRole;
 import santos.tiago.back_end.repository.UserRepository;
 import santos.tiago.back_end.service.UserService;
 
@@ -32,7 +33,12 @@ public class ApiController {
     }
 
     @GetMapping("/usuarios")
-    public ResponseEntity<List<UserResponse>> getUsers() {
+    public ResponseEntity<List<UserResponse>> getUsers(Authentication authentication) {
+        Optional<User> currentUser = userRepository.findByUsername(authentication.getName());
+        if (currentUser.isPresent() && currentUser.get().getRole() == UserRole.CLIENT) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized to read users.");
+        }
+
         List<UserResponse> target = new ArrayList<>();
         userRepository.findAll().forEach(
                 (user) -> target.add(
@@ -60,7 +66,10 @@ public class ApiController {
             }
         }
 
-        if (containsAuthority || user.get().getUsername().equals(authentication.getName())) {
+        Optional<User> currentUser = userRepository.findByUsername(authentication.getName());
+        boolean currentUserIsClient = currentUser.isPresent() && currentUser.get().getRole() == UserRole.CLIENT;
+
+        if ((containsAuthority && !currentUserIsClient) || user.get().getUsername().equals(authentication.getName())) {
             return new ResponseEntity<>(new UserResponse(user.get().getUsername(), user.get().getRole()), HttpStatus.OK);
         } else {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized to read data from user.");
@@ -68,10 +77,15 @@ public class ApiController {
     }
 
     @PostMapping("/usuarios")
-    public ResponseEntity<String> createUser(@NonNull @Valid @RequestBody User user) {
+    public ResponseEntity<String> createUser(@NonNull @Valid @RequestBody User user, Authentication authentication) {
         Optional<User> matchingUser = userRepository.findByUsername(user.getUsername());
         if (matchingUser.isPresent()) {
             throw new DuplicateKeyException("User already present in the database");
+        }
+
+        Optional<User> currentUser = userRepository.findByUsername(authentication.getName());
+        if (currentUser.isPresent() && currentUser.get().getRole() != UserRole.ADMIN) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized to create users.");
         }
 
         User newUser = this.userService.createUser(user);
@@ -80,14 +94,38 @@ public class ApiController {
     }
 
     @PutMapping("/usuarios/{username}")
-    public ResponseEntity<String> updateUser(@NonNull @PathVariable String username, @NonNull @Valid @RequestBody User user) {
+    public ResponseEntity<String> updateUser(@NonNull @PathVariable String username, @NonNull @Valid @RequestBody User user, Authentication authentication) {
+        Optional<User> userInRepository = userRepository.findByUsername(username);
+        if (userInRepository.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Username not found.");
+        }
+
         User userToUpdate = this.userService.createUser(user);
+
+        Optional<User> optionalCurrentUser = userRepository.findByUsername(authentication.getName());
+        if (optionalCurrentUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Current user not found.");
+        }
+
+        User currentUser = optionalCurrentUser.get();
+        if (currentUser.getRole() == UserRole.CLIENT) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized to update users.");
+        }
+        if (currentUser.getRole() == UserRole.OPERATOR && (userToUpdate.getRole() == UserRole.ADMIN || userInRepository.get().getRole() == UserRole.ADMIN)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized to set role higher than self.");
+        }
+
         this.userRepository.update(username, userToUpdate.getPassword(), userToUpdate.getRole());
         return new ResponseEntity<>("Usuário alterado.", HttpStatus.OK);
     }
 
     @DeleteMapping("/usuarios/{username}")
-    public ResponseEntity<String> deleteUser(@NonNull @PathVariable String username) {
+    public ResponseEntity<String> deleteUser(@NonNull @PathVariable String username, Authentication authentication) {
+        Optional<User> currentUser = userRepository.findByUsername(authentication.getName());
+        if (currentUser.isPresent() && currentUser.get().getRole() != UserRole.ADMIN) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized to delete users.");
+        }
+
         this.userRepository.deleteById(username);
         return new ResponseEntity<>("Usuário deletado.", HttpStatus.NO_CONTENT);
     }
